refactor(cli): extract BookCreateCommands from BookCommands
- Migrate book creation logic (`scan`, `registerBook`, `createBookManually`) to new `BookCreateCommands` class.
- Remove redundant methods and unused imports from `BookCommands`.
- Improve separation of concerns by isolating book creation-related commands into a dedicated class.
