+++ src/main/java/com/penrose/bibby/cli/command/book/BookCommands.java
@@ -72,220 +72,17 @@ public class BookCommands extends AbstractShellComponent {
     // ───────────────────────────────────────────────────────────────────
 
 
-    /**
-     * Scans a book's ISBN and facilitates its addition to the library database. This method
-     * allows users to interactively scan a book, validate its ISBN, retrieve metadata,
-     * and confirm or modify its placement in the library shelving system.
-     *
-     * @param multi if true, enables multiple book scanning functionality (currently unused);
-     *              if false, initiates a single book scanning process.
-     */
-    @Command(command = "scan", description = "Scan a book's ISBN to facilitate its addition to the library database.",group = "Book Commands")
-    public void scanBook(
-            @ShellOption (defaultValue = "single") boolean multi) {
-
-//        if (multi) multiBookScan();
-        log.info("Initiating scanBook for Single Scan.");
-        System.out.println("\n\u001B[95mSingle Book Scan");
-        String isbn = cliPrompt.promptForIsbnScan();
-        if (!cliPrompt.isbnValidator(isbn)) return;
-        BookMetaDataResponse bookMetaDataResponse = bookFacade.findBookMetaDataByIsbn(isbn);
-        log.debug("BookMetaDataResponse received: {}", bookMetaDataResponse);
 
 
-        log.debug("Authors verified/created for book.");
-        log.info(bookMetaDataResponse.toString());
-        String bookcard = bookcardRenderer.createBookCard(bookMetaDataResponse.title(),
-                bookMetaDataResponse.isbn(),
-                bookMetaDataResponse.authors().toString(),
-                bookMetaDataResponse.publisher(),
-                "PENDING / NOT SET",
-                "PENDING / NOT SET",
-                "PENDING / NOT SET");
-        System.out.println(bookcard);
-
-        if (cliPrompt.promptBookConfirmation()) {
-            String location = cliPrompt.promptForBookcaseLocation();
-            System.out.println("Selected Location: " + location);
-            Long bookcaseId = cliPrompt.promptForBookCase(promptOptions.bookCaseOptionsByLocation(location));
-            if(bookcaseId == null){
-                return;
-            }
-            Long shelfId = cliPrompt.promptForShelf(bookcaseId);
-            if(shelfId == null){
-                return;
-            }
-            List<Long> authorIds = createAuthorsFromMetaData(bookMetaDataResponse.authors());
-
-            bookFacade.createBookFromMetaData(bookMetaDataResponse, authorIds, isbn, shelfId);
-            String updatedBookCard = bookcardRenderer.createBookCard(bookMetaDataResponse.title(),
-                    bookMetaDataResponse.isbn(),
-                    bookMetaDataResponse.authors().toString(),
-                    bookMetaDataResponse.publisher(),
-                    bookcaseFacade.findBookCaseById(bookcaseId).get().bookcaseLabel(),
-                    shelfFacade.findShelfById(shelfId).get().shelfLabel(),
-                    bookcaseFacade.findBookCaseById(bookcaseId).get().location()
-
-                    );
-            System.out.println(updatedBookCard);
-        }
-
-    }
-
-    @Command(command = "new", description = "Create a new book entry")
-    public void registerBook(
-            @ShellOption(defaultValue = "false") boolean scan,
-            @ShellOption(defaultValue = "false") boolean multi) {
-
-        log.info("Starting new book registration process.");
-        log.debug("{} {}", scan, multi);
-
-        ScanMode mode = ScanMode.from(scan, multi);
-        switch(mode){
-            case SINGLE -> scanBook(false);
-//            case MULTI -> multiBookScan();
-            case NONE -> createBookManually();
-        }
-    }
 
 
-    public List<AuthorDTO> createAuthors(){
-        int numberOfAuthors = cliPrompt.promptForBookAuthorCount();
-        List<AuthorDTO> authors = new ArrayList<>();
-        for (int i = 0; i < numberOfAuthors; i++) {
-            AuthorDTO authorDTO = cliPrompt.promptForAuthor();
-            log.info("Collected Author Details: {}", authorDTO);
 
-            if(authorFacade.authorExistFirstNameLastName(authorDTO.firstName(),authorDTO.lastName())){
-                log.info("Author already exists: {} {}", authorDTO.firstName(), authorDTO.lastName());
-                System.out.println("Multiple Authors with this name.\n");
-                Long authorId = cliPrompt.promptMultipleAuthorConfirmation(authorDTO);
-                log.info("Existing author selected with ID: {}", authorId);
-                //todo: 0 is a magic number here, refactor needed
-                if(authorId == 0){
-                    log.info("Creating new author as per user request: {} {}", authorDTO.firstName(), authorDTO.lastName());
-                    authors.add(authorFacade.saveAuthor(authorDTO));
-                    log.info("Author saved: {}", authors.get(i));
-                }else{
-                    log.info("Fetching existing author with ID: {}", authorId);
 
-                    AuthorDTO existingAuthor = authorFacade.findById(authorId);
-                    authors.add(existingAuthor);
-                    log.info("Existing author added to list: {}", existingAuthor);
-                }
-            }else{
-                log.info("Creating new author: {} {}", authorDTO.firstName(), authorDTO.lastName());
-                authors.add(authorFacade.saveAuthor(authorDTO));
-                log.info("Author saved: {}", authors.get(i));
-            }
 
-        }
-        log.info("Authors Created and Returned: {}", authors);
-        return authors;
-    }
 
-    /**
-     * Checks if an author exists in the system with the given first and last name.
-     *
-     * @param firstName the first name of the author
-     * @param lastName the last name of the author
-     * @return true if an author with the specified first and last name exists, false otherwise
-     */
-    public boolean authorExists(String firstName, String lastName){
-        return authorFacade.authorExistFirstNameLastName(firstName,lastName);
-    }
 
-    /**
-     * Creates a new author if the given author does not exist in the system,
-     * or associates an existing author based on user input when multiple matches are found.
-     *
-     * @param authorDTO the data transfer object containing the first name and last name of the author
-     * @return the ID of the newly created or existing author, or null if no valid selection is made
-     */
-    public Long createNewAuthorOrAddExisting(AuthorDTO authorDTO){
-        String firstName = authorDTO.firstName();
-        String lastName = authorDTO.lastName();
 
-        if(authorExists(firstName,lastName)) {
-            log.info("Author already exists: {} {}", authorDTO.firstName(), authorDTO.lastName());
-            System.out.println("Multiple Authors with this name.\n");
-            Long authorId = cliPrompt.promptMultipleAuthorConfirmation(authorDTO);
-            log.info("Existing author selected with ID: {}", authorId);
-            //todo: 0 is a magic number here, refactor needed
-            if (authorId == 0) {
-                log.info("Creating new author as per user request: {} {}", authorDTO.firstName(), authorDTO.lastName());
-                return (authorFacade.saveAuthor(authorDTO).id());
-//                log.info("Author saved: {}", firstName + " " + lastName);
-            } else {
-                log.info("Fetching existing author with ID: {}", authorId);
 
-                AuthorDTO existingAuthor = authorFacade.findById(authorId);
-                return (existingAuthor.id());
-//                log.info("Existing author added to list: {}", existingAuthor);
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Saves a new author to the system database.
-     *
-     * @param authorDTO the data transfer object containing the author's details to be saved
-     * @return the AuthorDTO object corresponding to the newly saved author
-     */
-    public AuthorDTO saveNewAuthor(AuthorDTO authorDTO){
-        return authorFacade.saveAuthor(authorDTO);
-    }
-
-    /**
-     * Converts a given author name string into an AuthorDTO object.
-     * The full name is split into first and last name based on spaces.
-     * If the name contains only one word, it is treated as the first name, and the last name is left empty.
-     *
-     * @param authorName the string representing the author's name, expected in "FirstName LastName" format
-     * @return an AuthorDTO object with firstName and lastName extracted from the input string
-     */
-    public AuthorDTO mapAuthorStringToDTO(String authorName){
-        String[] nameParts = authorName.trim().split(" ");
-        String firstName = nameParts[0];
-        String lastName = nameParts.length > 1 ? nameParts[nameParts.length - 1] : "";
-        return new AuthorDTO(null, firstName, lastName);
-    }
-
-    /**
-     * Creates a list of author IDs by processing a list of author names.
-     * Each name is converted into an AuthorDTO, checked for existence,
-     * and either added as a new author or matched to an existing author.
-     *
-     * @param authorNames a list of strings representing author names, where each name
-     *                    is expected to be in a "FirstName LastName" format
-     * @return a list of author IDs corresponding to the processed authors
-     */
-    public List<Long> createAuthorsFromMetaData(List<String> authorNames){
-        List<Long> authors = new ArrayList<>();
-
-        for (String authorName : authorNames) {
-            AuthorDTO authorDTO = mapAuthorStringToDTO(authorName);
-            boolean exists = authorExists(authorDTO.firstName(), authorDTO.lastName());
-            authors.add(exists ? createNewAuthorOrAddExisting(authorDTO)
-                    : saveNewAuthor(authorDTO).id());
-        }
-        return authors;
-    }
-
-    public void createBookManually() {
-        System.out.println("\n\u001B[95mCreate New Book");
-
-        String title = cliPrompt.promptForBookTitle();
-        List<AuthorDTO> authors = createAuthors();
-        String isbn = cliPrompt.promptForBookIsbn();
-        log.info("Collected Book Details - Title: {}, ISBN: {}, Authors: {}", title, isbn, authors);
-
-        BookRequestDTO bookRequestDTO = new BookRequestDTO(title,isbn,authors);
-        log.info("BookRequestDTO created: {}", bookRequestDTO);
-
-        bookFacade.createNewBook(bookRequestDTO);
-    }
 
     @Command(command = "shelf", description = "Place a book on a shelf or move it to a new location.")
     public void addToShelf(){
